backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
finvizfinance==1.3.0
python-dotenv==1.2.1
requests==2.32.5
newspaper3k
lxml
deep-translator

